Add back mute, move announcer code into own file
