+++ src/keymap.tsx
@@ -3,8 +3,9 @@ import CodeMirror, { Editor } from 'codemirror';
 import SHARED from './shared';
 import {delete_, copy, paste, InsertTarget,
   ReplaceNodeTarget, OverwriteTarget, activateByNid} from './actions';
-import {partition, getRoot, skipCollapsed, say, mac,
+import {partition, getRoot, skipCollapsed, mac,
   getLastVisibleNode, preambleUndoRedo, playSound, BEEP} from './utils';
+import { say } from "./announcer";
 import {findAdjacentDropTargetId as getDTid} from './components/DropTarget';
 
 import type { AppDispatch } from './store';
