+++ src/ui/Search.tsx
@@ -2,7 +2,8 @@ import React, {Component} from 'react';
 import Dialog from '../components/Dialog';
 import {Tab, Tabs, TabList, TabPanel} from 'react-tabs';
 import 'react-tabs/style/react-tabs.less';
-import {say, getBeginCursor, getEndCursor, playSound, WRAP} from '../utils';
+import {getBeginCursor, getEndCursor, playSound, WRAP} from '../utils';
+import { say } from "../announcer";
 import { BlockEditorComponentClass, Search } from './BlockEditor';
 import { Searcher } from './searchers/Searcher';
 import { GetProps } from 'react-redux';
